+++ subiquity/server/controllers/__init__.py
@@ -16,6 +16,7 @@
 from .cmdlist import EarlyController, LateController, ErrorController
 from .debconf import DebconfController
 from .filesystem import FilesystemController
+from .identity import IdentityController
 from .install import InstallController
 from .keyboard import KeyboardController
 from .locale import LocaleController
@@ -33,6 +34,7 @@ __all__ = [
     'EarlyController',
     'ErrorController',
     'FilesystemController',
+    'IdentityController',
     'InstallController',
     'KeyboardController',
     'LateController',
